automation-system/llm1-diagnostics/requirements.txt
flask==2.2.5
requests==2.31.0
python-dotenv==0.21.0
openai==1.3.0
